requirements.txt
anyio==3.6.1
asttokens==2.0.8
backcall==0.2.0
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.1.1
fastapi==0.85.0
flake8==5.0.4
fonttools==4.37.4
greenlet==1.1.3.post0
h11==0.14.0
httptools==0.5.0
idna==3.4
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
joblib==1.2.0
jupyter-client==7.3.5
jupyter-core==4.11.1
kiwisolver==1.4.4
matplotlib==3.6.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==0.4.3
nest-asyncio==1.5.6
numpy==1.23.3
packaging==21.3
pandas==1.5.0
pandasql==0.7.3
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pydantic==1.10.2
pyflakes==2.5.0
Pygments==2.13.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.4
PyYAML==6.0
pyzmq==24.0.1
scikit-learn==1.1.2
scipy==1.9.2
seaborn==0.12.0
six==1.16.0
sklearn==0.0
sniffio==1.3.0
SQLAlchemy==1.4.41
stack-data==0.5.1
starlette==0.20.4
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.2
traitlets==5.4.0
typing-extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
wcwidth==0.2.5
websockets==10.3
